refactor : useProduct mutation 코드 추가
